pyproject.toml
[project]
name = "scoutapp-program-builder"
version = "0.1.0"
description = "Una aplicación de chat inteligente para crear programas Scout."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
"flet==0.28.3",
"openai>=1.0",
"python-dotenv>=1.0",
"pypdf>=4.0",
"sentence-transformers>=2.0",
"faiss-cpu>=1.7"
]


[[project.authors]]
name = "Joaquín Eduardo Ramos Farfán"
email = "contacto@joaquinramos.dev"


[tool.flet]
org = "com.joaquinramos"
product = "scoutapp-program-builder"
company = "Joaquin Ramos Developer"
copyright = "Copyright (C) 2025 by Joaquín Eduardo Ramos Farfán"


[tool.flet.app]
path = "src"


[tool.uv]
dev-dependencies = [ "flet[all]==0.28.3" ]


[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies.flet]
extras = [ "all" ]
version = "0.28.3"
